+++ libs/rxjs/number/src/lib/to-locale-string.ts
@@ -11,7 +11,11 @@ import { map } from 'rxjs/operators';
  * [Number.prototype.toLocaleString](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number/toLocaleString)
  * and the passed locale option.
  *
+ * The `toLocaleString` uses locale options in [Intl.NumberFormat](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Intl/NumberFormat)
+ * to provide additional formatting for values types such as currency
+ *
  * @param locales The locale or locales that the number is being formatted for
+ * @param format Formatting of the string based on [Intl.NumberFormat](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Intl/NumberFormat)
  *
  * @example
  * ```ts
@@ -26,17 +30,6 @@ import { map } from 'rxjs/operators';
  *  .subscribe(console.log) // '1,000,000'
  * ```
  *
- * @returns String of the number formatted using `Number.prototype.toLocaleString`
- * @category RxJS Number Formatting
- */
-function toLocaleString(locales: string | string[]): OperatorFunction<number, string>;
-/**
- * The `toLocaleString` uses locale options in [Intl.NumberFormat](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Intl/NumberFormat)
- * to provide additional formatting for values types such as currency
- *
- * @param locales The locale or locales that the number is being formatted for
- * @param format Formatting of the string based on [Intl.NumberFormat](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Intl/NumberFormat)
- *
  * @example
  * ```ts
  * fromNumber(1000000)
@@ -47,12 +40,9 @@ function toLocaleString(locales: string | string[]): OperatorFunction<number, st
  * @returns String of the number formatted using `Number.prototype.toLocaleString`
  * @category RxJS Number Formatting
  */
-function toLocaleString(locales: string | string[], format: Intl.NumberFormatOptions): OperatorFunction<number, string>;
-function toLocaleString(
+export function toLocaleString(
   locales: string | string[],
   format?: Intl.NumberFormatOptions,
 ): OperatorFunction<number, string> {
   return (source: Observable<number>) => source.pipe(map((number) => number.toLocaleString(locales, format)));
 }
-
-export { toLocaleString };