+++ libs/rxjs/number/src/lib/to-string.ts
@@ -12,22 +12,16 @@ import { map } from 'rxjs/operators';
  * The operator will return a string value using
  * [Number.prototype.toString](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number/toString)
  *
+ * A number can be formatted to a different base, such as base2 (boolean) or * base16 (hexadecimal)
+ *
+ * @param radix The base number to format to. Default is 10
+ *
  * @example
  * ```ts
  * fromNumber(42)
  *  .pipe(toString())
  *  .subscribe(console.log) // '42'
  * ```
- *
- * @returns A number that is parsed from a string using `Number.parseInt`
- * @category RxJS Number Parsing
- */
-function toString(): OperatorFunction<number, string>;
-/**
- * A number can be formatted to a different base, such as base2 (boolean) or
- * base16 (hexadecimal)
- *
- * @param radix The base number to format to. Default is 10
  *
  * @example
  * ```ts
@@ -39,8 +33,6 @@ function toString(): OperatorFunction<number, string>;
  * @returns A number that is parsed from a string using `Number.parseInt` with radix
  * @category RxJS Number Parsing
  */
-function toString(radix: number): OperatorFunction<number, string>;
-function toString(radix = 10): OperatorFunction<number, string> {
+export function toString(radix = 10): OperatorFunction<number, string> {
   return (source: Observable<number>) => source.pipe(map((number) => number.toString(radix)));
 }
-export { toString };